Build native messaging host for macOS
